requirements.txt
# requirements.txt
streamlit
pandas
matplotlib
geopandas
folium
streamlit-folium
google-generativeai